Implement bi-directional batch fetching
To prevent confusion with other directional types and reduce confusion
when dealing with RTL and inverted collections, we are using the term
"batch fetching dimension" and the values of "head" and "tail". By
default a collection view only supports "tail" fetching for backwards
compatibility (and since that's the common case).